Fix bugs, let the server finally run

+++ TDS_Server/Instance/Lobby/FightLobby/Weapon.cs
@@ -10,9 +10,9 @@ namespace TDS_Server.Instance.Lobby
             player.RemoveAllWeapons();
             foreach (LobbyWeapons weapon in LobbyEntity.LobbyWeapons)
             {
-                if (!System.Enum.IsDefined(typeof(WeaponHash), weapon.Hash))
+                if (!System.Enum.IsDefined(typeof(WeaponHash), (uint) weapon.Hash))
                     continue;
-                WeaponHash hash = (WeaponHash)weapon.Hash;
+                WeaponHash hash = (WeaponHash) ((uint)weapon.Hash);
                 player.GiveWeapon(hash, 0);
                 player.SetWeaponAmmo(hash, (int)weapon.Ammo);
             }
